fixed the codechecker errors outlined in the travis ci report and removed the copyright handles that were above the functions
